requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.2
django-appconf==1.0.6
django-rest-framework==0.1.0
django-webauthn==0.0.2
django-webpush==0.3.6
djangorestframework==3.15.2
frozenlist==1.5.0
google-auth==2.35.0
http_ece==1.2.1
idna==3.10
multidict==6.1.0
propcache==0.2.0
py-vapid==1.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyfcm==2.0.7
pywebpush==2.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
yarl==1.17.1
